add channel list functionality to pass tests, remove channel container and merge into channel list
